Add prng module; move XorShiftRng to it
